Split createFormMachine into smaller utils
