+++ src/forms/createFormMachine.ts
@@ -1,55 +1,10 @@
 import { assign, setup } from "xstate";
-import type { FormData } from "@/constants/fields";
-import type { FormConfig } from "@/constants/forms";
 import type {
   FormMachineContext,
   FormMachineEvent,
   FormPhase,
   Step,
 } from "./types";
-
-/**
- * DSL helper: marks a step for inclusion in a form flow.
- */
-export function step(config: Step): Step {
-  return config;
-}
-
-/**
- * Finds the next step index (forward) whose guard passes, starting after `fromIndex`.
- * Returns -1 if no eligible step exists (meaning we should advance to review).
- */
-export function findNextStepIndex(
-  steps: readonly Step[],
-  fromIndex: number,
-  formData: Partial<FormData>,
-): number {
-  for (let i = fromIndex + 1; i < steps.length; i++) {
-    const s = steps[i];
-    if (!s.guard || s.guard(formData)) {
-      return i;
-    }
-  }
-  return -1;
-}
-
-/**
- * Finds the previous step index (backward) whose guard passes, starting before `fromIndex`.
- * Returns -1 if no eligible step exists (meaning we should go back to title).
- */
-export function findPrevStepIndex(
-  steps: readonly Step[],
-  fromIndex: number,
-  formData: Partial<FormData>,
-): number {
-  for (let i = fromIndex - 1; i >= 0; i--) {
-    const s = steps[i];
-    if (!s.guard || s.guard(formData)) {
-      return i;
-    }
-  }
-  return -1;
-}
 
 interface CreateFormMachineOptions {
   id: string;
@@ -150,38 +105,4 @@ export function getPhase(stateValue: unknown): FormPhase {
     return Object.keys(stateValue)[0] as FormPhase;
   }
   return "title";
-}
-
-/**
- * Gets the ordered list of visible step IDs given the current form data and steps.
- */
-export function getVisibleStepIds(
-  steps: readonly Step[],
-  formData: Partial<FormData>,
-): string[] {
-  return steps.filter((s) => !s.guard || s.guard(formData)).map((s) => s.id);
-}
-
-/**
- * Extracts and flattens all field names from a steps array.
- */
-export function fieldsFromSteps(steps: readonly Step[]) {
-  return steps.flatMap((s) => s.fields);
-}
-
-/**
- * Creates a complete FormConfig from the unique per-form properties.
- * Derives `machine` and `fields` from `slug` and `steps`.
- */
-export function defineFormConfig(
-  input: Omit<FormConfig, "machine" | "fields">,
-): FormConfig {
-  const { slug, steps, ...rest } = input;
-  return {
-    slug,
-    steps,
-    machine: createFormMachine({ id: slug, steps }),
-    fields: fieldsFromSteps(steps),
-    ...rest,
-  };
 }
