+++ src/pages/forms/court-order-ma/config.ts
@@ -1,4 +1,4 @@
-import { defineFormConfig, step } from "@/forms/createFormMachine";
+import { defineFormConfig, step } from "@/forms/defineFormConfig";
 import { addressStep } from "./_steps/AddressStep";
 import { birthplaceStep } from "./_steps/BirthplaceStep";
 import { contactInfoStep } from "./_steps/ContactInfoStep";
